% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{compute.day.length}
\alias{compute.day.length}
\title{compute.day.length}
\usage{
compute.day.length(doy, lat.deg)
}
\arguments{
\item{lat.deg}{}
}
\value{
day length (h)
}
\description{
compute.day.length
}

