% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{compute.et0.Hargreaves}
\alias{compute.et0.Hargreaves}
\title{compute.et0.Hargreaves}
\usage{
compute.et0.Hargreaves(doy, lat.deg, tmin.C, tmax.C, tave.C = NULL)
}
\arguments{
\item{tave.C}{}
}
\value{
ET0 (mm)
}
\description{
compute.et0.Hargreaves
}

