% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{compute.extraterrestrial.radiation}
\alias{compute.extraterrestrial.radiation}
\title{compute.extraterrestrial.radiation}
\usage{
compute.extraterrestrial.radiation(doy, lat.deg)
}
\arguments{
\item{lat.deg}{}
}
\value{
Ra: extraterrestrial.radiation
}
\description{
compute.extraterrestrial.radiation
}

