% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{compute.incoming.solar.radiation.Hargreaves}
\alias{compute.incoming.solar.radiation.Hargreaves}
\title{compute.incoming.solar.radiation.Hargreaves}
\usage{
compute.incoming.solar.radiation.Hargreaves(doy, lat.deg, tmin.C, tmax.C,
  krs = 0.19)
}
\arguments{
\item{krs}{}
}
\value{
incoming solar radiation (MJ m-2 d-1)
}
\description{
compute.incoming.solar.radiation.Hargreaves
}

